Personal/lunk alarm/requirements.txt
SpeechRecognition==3.8.1
textblob==0.15.3
RPi.GPIO==0.7.0